+++ src/main/java/imb/pr3/estetica/controller/APIResponse.java
@@ -1,0 +1,2 @@
+package imb.pr3.estetica.controller;public class APIResponse {
+}
