+++ src/main/java/imb/pr3/estetica/service/IMetodoDePagoService.java
@@ -1,4 +1,4 @@
-package imb.pr3.estetica.servicio;
+package imb.pr3.estetica.service;
 
 import java.util.List;
